requirements.txt
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
google-generativeai>=0.3.0
qdrant-client>=1.6.0
numpy>=1.21.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
flask==2.0.1
google-api-python-client==2.0.0
google-auth-oauthlib==0.4.6
google-generativeai==0.3.1
qdrant-client==1.1.1
python-dotenv==0.19.0
numpy==1.21.0
gunicorn==20.1.0
python-dotenv==0.19.0